--- a/gin_integration_test.go
+++ b/gin_integration_test.go
@@ -22,7 +22,14 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func testRequest(t *testing.T, url string) {
+// params[0]=url example:http://127.0.0.1:8080/index (cannot be empty)
+// params[1]=response body (custom compare content)
+func testRequest(t *testing.T, params ...string) {
+
+	if len(params) == 0 {
+		t.Fatal("url cannot be empty")
+	}
+
 	tr := &http.Transport{
 		TLSClientConfig: &tls.Config{
 			InsecureSkipVerify: true,
@@ -30,13 +37,18 @@ func testRequest(t *testing.T, url string) {
 	}
 	client := &http.Client{Transport: tr}
 
-	resp, err := client.Get(url)
+	resp, err := client.Get(params[0])
 	assert.NoError(t, err)
 	defer resp.Body.Close()
 
 	body, ioerr := ioutil.ReadAll(resp.Body)
 	assert.NoError(t, ioerr)
-	assert.Equal(t, "it worked", string(body), "resp body should match")
+
+	var expected = "it worked"
+	if len(params) > 1 {
+		expected = params[1]
+	}
+	assert.Equal(t, expected, string(body), "resp body should match")
 	assert.Equal(t, "200 OK", resp.Status, "should get a 200")
 }
 
@@ -372,4 +384,34 @@ func testGetRequestHandler(t *testing.T, h http.Handler, url string) {
 
 	assert.Equal(t, "it worked", w.Body.String(), "resp body should match")
 	assert.Equal(t, 200, w.Code, "should get a 200")
+}
+
+func TestRunDynamicRouting(t *testing.T) {
+	router := New()
+	router.GET("/aa/*xx", func(c *Context) { c.String(http.StatusOK, "/aa/*xx") })
+	router.GET("/ab/*xx", func(c *Context) { c.String(http.StatusOK, "/ab/*xx") })
+	router.GET("/", func(c *Context) { c.String(http.StatusOK, "home") })
+	router.GET("/:cc", func(c *Context) { c.String(http.StatusOK, "/:cc") })
+	router.GET("/:cc/cc", func(c *Context) { c.String(http.StatusOK, "/:cc/cc") })
+	router.GET("/get/test/abc/", func(c *Context) { c.String(http.StatusOK, "/get/test/abc/") })
+	router.GET("/get/:param/abc/", func(c *Context) { c.String(http.StatusOK, "/get/:param/abc/") })
+
+	ts := httptest.NewServer(router)
+	defer ts.Close()
+
+	testRequest(t, ts.URL+"/", "home")
+	testRequest(t, ts.URL+"/aa/aa", "/aa/*xx")
+	testRequest(t, ts.URL+"/ab/ab", "/ab/*xx")
+	testRequest(t, ts.URL+"/all", "/:cc")
+	testRequest(t, ts.URL+"/all/cc", "/:cc/cc")
+	testRequest(t, ts.URL+"/a/cc", "/:cc/cc")
+	testRequest(t, ts.URL+"/a", "/:cc")
+	testRequest(t, ts.URL+"/get/test/abc/", "/get/test/abc/")
+	testRequest(t, ts.URL+"/get/te/abc/", "/get/:param/abc/")
+	testRequest(t, ts.URL+"/get/xx/abc/", "/get/:param/abc/")
+	testRequest(t, ts.URL+"/get/tt/abc/", "/get/:param/abc/")
+	testRequest(t, ts.URL+"/get/a/abc/", "/get/:param/abc/")
+	testRequest(t, ts.URL+"/get/t/abc/", "/get/:param/abc/")
+	testRequest(t, ts.URL+"/get/aa/abc/", "/get/:param/abc/")
+	testRequest(t, ts.URL+"/get/abas/abc/", "/get/:param/abc/")
 }