--- a/tree.go
+++ b/tree.go
@@ -600,7 +600,8 @@ walk: // Outer loop for walking the tree
 			return
 		}
 
-		if path != "/" && skippedPath != "" {
+		// path != "/" && skippedPath != ""
+		if len(path) != 1 && len(skippedPath) > 0 && strings.HasSuffix(skippedPath, path) {
 			path = skippedPath
 			n = latestNode
 			skippedPath = ""