@appleboy 2021-06-28
Please also add benchmark result like https://github.com/gin-gonic/gin/pull/2706#issuecomment-828915618 comment.