@qm012 2021-06-28
> Please also add benchmark result like [#2706 (comment)](https://github.com/gin-gonic/gin/pull/2706#issuecomment-828915618) comment.

hi，local 'go test' passed and add benchmark if there are omissions, please help me to add. 😆

The previous code could not match when running 'benchmark', and 404 not found occurred,so I fix the code.😆

The benchmarks:

**this PR** - https://travis-ci.com/github/qm012/go-http-routing-benchmark/builds/231071258

The `benchstat` report:

```shell
BenchmarkGin_Param        	53810290	        68.2 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_Param5       	24983499	       144 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_Param20      	 8174235	       432 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_ParamWrite   	27743445	       129 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_GithubStatic 	28786916	       125 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_GithubParam  	15244320	       236 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_GithubAll    	   83005	     43175 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_GPlusStatic  	43877554	        82.3 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_GPlusParam   	31455780	       114 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_GPlus2Params 	23434296	       154 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_GPlusAll     	 2101560	      1695 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_ParseStatic  	44035694	        81.7 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_ParseParam   	43929930	        82.0 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_Parse2Params 	34703761	       104 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_ParseAll     	 1215774	      2948 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_StaticAll    	   97797	     36469 ns/op	       0 B/op	       0 allocs/op
```

Thank you all for your reviews.