@qm012 2021-06-29
**this PR Benchmark** https://travis-ci.com/github/qm012/go-http-routing-benchmark/builds/231163846
commit: [4e8e2cb](https://github.com/qm012/gin/tree/develop-2762)

```shell
goos: linux
goarch: amd64
pkg: github.com/julienschmidt/go-http-routing-benchmark
BenchmarkGin_Param        	54527260	        66.8 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_Param5       	25176636	       143 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_Param20      	 8204526	       431 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_ParamWrite   	28036678	       128 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_GithubStatic 	29023458	       126 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_GithubParam  	15394497	       233 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_GithubAll    	   84313	     42196 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_GPlusStatic  	44162476	        83.2 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_GPlusParam   	30823569	       115 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_GPlus2Params 	23545215	       155 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_GPlusAll     	 2134837	      1663 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_ParseStatic  	44415607	        83.2 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_ParseParam   	43003294	        81.9 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_Parse2Params 	34613594	       105 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_ParseAll     	 1238673	      2868 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_StaticAll    	   99607	     35951 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/julienschmidt/go-http-routing-benchmark	65.859s
```

**Local Benchmark**
```shell
goos: windows
goarch: amd64
pkg: github.com/julienschmidt/go-http-routing-benchmark
BenchmarkGin_Param              28237021               112 ns/op               0 B/op          0 allocs/op
BenchmarkGin_Param5             15618274               233 ns/op               0 B/op          0 allocs/op
BenchmarkGin_Param20             5198558               671 ns/op               0 B/op          0 allocs/op
BenchmarkGin_ParamWrite         17562182               204 ns/op               0 B/op          0 allocs/op
BenchmarkGin_GithubStatic       17266526               206 ns/op               0 B/op          0 allocs/op
BenchmarkGin_GithubParam         8999959               403 ns/op               0 B/op          0 allocs/op
BenchmarkGin_GithubAll             52173             69864 ns/op               0 B/op          0 allocs/op
BenchmarkGin_GPlusStatic        26471035               133 ns/op               0 B/op          0 allocs/op
BenchmarkGin_GPlusParam         18701162               192 ns/op               0 B/op          0 allocs/op
BenchmarkGin_GPlus2Params       13284082               262 ns/op               0 B/op          0 allocs/op
BenchmarkGin_GPlusAll            1357976              2629 ns/op               0 B/op          0 allocs/op
BenchmarkGin_ParseStatic        25174754               135 ns/op               0 B/op          0 allocs/op
BenchmarkGin_ParseParam         26181741               146 ns/op               0 B/op          0 allocs/op
BenchmarkGin_Parse2Params       20690522               173 ns/op               0 B/op          0 allocs/op
BenchmarkGin_ParseAll             757886              4629 ns/op               0 B/op          0 allocs/op
BenchmarkGin_StaticAll             59404             55964 ns/op               0 B/op          0 allocs/op
PASS
ok      github.com/julienschmidt/go-http-routing-benchmark      66.332s
```