@qm012 2021-06-29
The benchmarks:

**master** https://travis-ci.com/github/qm012/go-http-routing-benchmark/builds/231184203
**this PR** https://travis-ci.com/github/qm012/go-http-routing-benchmark/builds/231184328

The benchstat report:
**master**
```shell
BenchmarkGin_Param        	50492419	        72.0 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_Param5       	29414568	       122 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_Param20      	11018823	       320 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_ParamWrite   	28458553	       127 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_GithubStatic 	39511770	        91.2 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_GithubParam  	22447558	       160 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_GithubAll    	  112528	     31926 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_GPlusStatic  	55041181	        65.4 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_GPlusParam   	38489194	        93.7 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_GPlus2Params 	30157081	       119 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_GPlusAll     	 2792388	      1289 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_ParseStatic  	54590468	        66.0 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_ParseParam   	47282571	        75.8 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_Parse2Params 	41202753	        87.5 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_ParseAll     	 1616793	      2225 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_StaticAll    	  150019	     24104 ns/op	       0 B/op	       0 allocs/op
```
**this PR**
```shell
BenchmarkGin_Param        	49536423	        73.4 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_Param5       	24659034	       146 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_Param20      	 8172079	       432 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_ParamWrite   	27776470	       130 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_GithubStatic 	27761433	       130 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_GithubParam  	15148544	       238 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_GithubAll    	   83085	     48136 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_GPlusStatic  	38024416	        95.3 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_GPlusParam   	28010620	       137 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_GPlus2Params 	23566668	       153 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_GPlusAll     	 2082960	      1726 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_ParseStatic  	42094258	        85.2 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_ParseParam   	41519977	        86.4 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_Parse2Params 	34881759	       103 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_ParseAll     	 1234374	      2920 ns/op	       0 B/op	       0 allocs/op
BenchmarkGin_StaticAll    	   99387	     36161 ns/op	       0 B/op	       0 allocs/op
```