@appleboy 2021-06-29
@qm012 I make the PR https://github.com/qm012/go-http-routing-benchmark/pull/1/files to change the go version and update the testing count. Please update the new result by `benchstat` tool below.

```sh
$ benchstat old.txt new.txt
name              old time/op    new time/op    delta
Gin_Param           72.0ns ± 0%    73.4ns ± 0%   ~     (p=1.000 n=1+1)
Gin_Param5           122ns ± 0%     146ns ± 0%   ~     (p=1.000 n=1+1)
Gin_Param20          320ns ± 0%     432ns ± 0%   ~     (p=1.000 n=1+1)
Gin_ParamWrite       127ns ± 0%     130ns ± 0%   ~     (p=1.000 n=1+1)
Gin_GithubStatic    91.2ns ± 0%   130.0ns ± 0%   ~     (p=1.000 n=1+1)
Gin_GithubParam      160ns ± 0%     238ns ± 0%   ~     (p=1.000 n=1+1)
Gin_GithubAll       31.9µs ± 0%    48.1µs ± 0%   ~     (p=1.000 n=1+1)
Gin_GPlusStatic     65.4ns ± 0%    95.3ns ± 0%   ~     (p=1.000 n=1+1)
Gin_GPlusParam      93.7ns ± 0%   137.0ns ± 0%   ~     (p=1.000 n=1+1)
Gin_GPlus2Params     119ns ± 0%     153ns ± 0%   ~     (p=1.000 n=1+1)
Gin_GPlusAll        1.29µs ± 0%    1.73µs ± 0%   ~     (p=1.000 n=1+1)
Gin_ParseStatic     66.0ns ± 0%    85.2ns ± 0%   ~     (p=1.000 n=1+1)
Gin_ParseParam      75.8ns ± 0%    86.4ns ± 0%   ~     (p=1.000 n=1+1)
Gin_Parse2Params    87.5ns ± 0%   103.0ns ± 0%   ~     (p=1.000 n=1+1)
Gin_ParseAll        2.23µs ± 0%    2.92µs ± 0%   ~     (p=1.000 n=1+1)
Gin_StaticAll       24.1µs ± 0%    36.2µs ± 0%   ~     (p=1.000 n=1+1)
```