@qm012 2021-06-29
hi, @appleboy Thanks you for reminding me.
The benchmarks:

**master** https://travis-ci.com/github/qm012/go-http-routing-benchmark/builds/231194542
**this PR** https://travis-ci.com/github/qm012/go-http-routing-benchmark/builds/231194421

The `benchstat` report:
```shell
$ benchstat old.txt new.txt
name              old time/op    new time/op    delta
Gin_Param           91.1ns ± 2%   119.2ns ± 4%  +30.92%  (p=0.000 n=10+10)
Gin_Param5           163ns ± 2%     253ns ± 3%  +54.70%  (p=0.000 n=9+10)
Gin_Param20          399ns ± 2%     712ns ± 8%  +78.57%  (p=0.000 n=9+10)
Gin_ParamWrite       174ns ± 1%     219ns ± 5%  +26.21%  (p=0.000 n=10+9)
Gin_GithubStatic     129ns ± 1%     223ns ± 1%  +72.08%  (p=0.000 n=9+9)
Gin_GithubParam      229ns ± 2%     429ns ± 2%  +87.68%  (p=0.000 n=10+9)
Gin_GithubAll       42.8µs ± 2%    73.0µs ± 2%  +70.56%  (p=0.000 n=9+10)
Gin_GPlusStatic     91.5ns ± 2%   147.0ns ± 2%  +60.60%  (p=0.000 n=10+9)
Gin_GPlusParam       137ns ± 2%     209ns ± 1%  +53.05%  (p=0.000 n=10+8)
Gin_GPlus2Params     180ns ± 1%     286ns ± 2%  +59.00%  (p=0.000 n=9+9)
Gin_GPlusAll        1.82µs ± 2%    2.86µs ± 1%  +57.30%  (p=0.000 n=10+10)
Gin_ParseStatic     99.3ns ± 3%   146.9ns ± 4%  +47.92%  (p=0.000 n=10+10)
Gin_ParseParam       106ns ± 1%     143ns ± 1%  +35.59%  (p=0.000 n=9+8)
Gin_Parse2Params     127ns ± 2%     183ns ± 3%  +43.87%  (p=0.000 n=10+10)
Gin_ParseAll        3.15µs ± 2%    4.90µs ± 2%  +55.56%  (p=0.000 n=10+10)
Gin_StaticAll       31.6µs ± 1%    59.4µs ± 2%  +87.94%  (p=0.000 n=9+10)

name              old alloc/op   new alloc/op   delta
Gin_Param            0.00B          0.00B          ~     (all equal)
Gin_Param5           0.00B          0.00B          ~     (all equal)
Gin_Param20          0.00B          0.00B          ~     (all equal)
Gin_ParamWrite       0.00B          0.00B          ~     (all equal)
Gin_GithubStatic     0.00B          0.00B          ~     (all equal)
Gin_GithubParam      0.00B          0.00B          ~     (all equal)
Gin_GithubAll        0.00B          0.00B          ~     (all equal)
Gin_GPlusStatic      0.00B          0.00B          ~     (all equal)
Gin_GPlusParam       0.00B          0.00B          ~     (all equal)
Gin_GPlus2Params     0.00B          0.00B          ~     (all equal)
Gin_GPlusAll         0.00B          0.00B          ~     (all equal)
Gin_ParseStatic      0.00B          0.00B          ~     (all equal)
Gin_ParseParam       0.00B          0.00B          ~     (all equal)
Gin_Parse2Params     0.00B          0.00B          ~     (all equal)
Gin_ParseAll         0.00B          0.00B          ~     (all equal)
Gin_StaticAll        0.00B          0.00B          ~     (all equal)

name              old allocs/op  new allocs/op  delta
Gin_Param             0.00           0.00          ~     (all equal)
Gin_Param5            0.00           0.00          ~     (all equal)
Gin_Param20           0.00           0.00          ~     (all equal)
Gin_ParamWrite        0.00           0.00          ~     (all equal)
Gin_GithubStatic      0.00           0.00          ~     (all equal)
Gin_GithubParam       0.00           0.00          ~     (all equal)
Gin_GithubAll         0.00           0.00          ~     (all equal)
Gin_GPlusStatic       0.00           0.00          ~     (all equal)
Gin_GPlusParam        0.00           0.00          ~     (all equal)
Gin_GPlus2Params      0.00           0.00          ~     (all equal)
Gin_GPlusAll          0.00           0.00          ~     (all equal)
Gin_ParseStatic       0.00           0.00          ~     (all equal)
Gin_ParseParam        0.00           0.00          ~     (all equal)
Gin_Parse2Params      0.00           0.00          ~     (all equal)
Gin_ParseAll          0.00           0.00          ~     (all equal)
Gin_StaticAll         0.00           0.00          ~     (all equal)
```