@qm012 2021-06-29
Benchstat report of the newly submitted Code:
```shell
$ benchstat old.txt new.txt
name              old time/op    new time/op    delta
Gin_Param           93.5ns ± 5%    99.5ns ± 0%   +6.49%  (p=0.000 n=10+7)
Gin_Param5           166ns ± 2%     216ns ± 1%  +30.03%  (p=0.000 n=10+9)
Gin_Param20          407ns ± 2%     612ns ± 1%  +50.42%  (p=0.000 n=10+9)
Gin_ParamWrite       177ns ± 2%     187ns ± 2%   +5.51%  (p=0.000 n=9+9)
Gin_GithubStatic     135ns ± 1%     193ns ± 2%  +43.04%  (p=0.000 n=9+10)
Gin_GithubParam      236ns ± 5%     367ns ± 1%  +55.40%  (p=0.000 n=10+9)
Gin_GithubAll       43.8µs ± 2%    62.1µs ± 3%  +41.81%  (p=0.000 n=10+9)
Gin_GPlusStatic     93.9ns ± 3%   123.6ns ± 2%  +31.58%  (p=0.000 n=9+9)
Gin_GPlusParam       139ns ± 1%     184ns ± 2%  +31.82%  (p=0.000 n=9+9)
Gin_GPlus2Params     183ns ± 1%     246ns ± 2%  +34.41%  (p=0.000 n=10+8)
Gin_GPlusAll        1.84µs ± 2%    2.49µs ± 2%  +35.36%  (p=0.000 n=10+9)
Gin_ParseStatic     93.2ns ± 1%   131.3ns ± 1%  +40.90%  (p=0.000 n=10+9)
Gin_ParseParam       108ns ± 2%     129ns ± 1%  +20.13%  (p=0.000 n=10+8)
Gin_Parse2Params     133ns ± 3%     160ns ± 1%  +20.37%  (p=0.000 n=9+9)
Gin_ParseAll        3.11µs ± 1%    4.32µs ± 2%  +38.70%  (p=0.000 n=8+10)
Gin_StaticAll       32.3µs ± 2%    51.2µs ± 1%  +58.16%  (p=0.000 n=10+10)
```
I will continue to optimize the report, but the current result is not very ideal
Root: Search the route tree for secondary route matching rules
```go
// match '/', If this condition is matched, the next route is found
if strings.HasSuffix(n.path, "/") || strings.Contains(n.fullPath, ":") || n.path == "" {
	matchNum++
}
```
Thank you all for your reviews.😆