@qm012 2021-06-30
The benchmarks:
**master** https://travis-ci.com/github/qm012/go-http-routing-benchmark/builds/231194542
**this PR** https://travis-ci.com/github/qm012/go-http-routing-benchmark/builds/231313687

The `benchstat` report:
```shell
name              old time/op    new time/op    delta
Gin_Param           62.5ns ± 4%    64.0ns ± 5%     ~     (p=0.110 n=10+10)
Gin_Param5           110ns ± 3%     114ns ± 5%   +3.31%  (p=0.017 n=10+10)
Gin_Param20          276ns ± 2%     298ns ± 3%   +7.92%  (p=0.000 n=10+10)
Gin_ParamWrite       118ns ± 2%     120ns ± 5%     ~     (p=0.136 n=10+10)
Gin_GithubStatic    78.3ns ± 2%    96.4ns ± 2%  +23.07%  (p=0.000 n=10+10)
Gin_GithubParam      129ns ± 0%     177ns ± 2%  +37.55%  (p=0.000 n=10+10)
Gin_GithubAll       25.5µs ± 0%    32.1µs ± 1%  +26.01%  (p=0.000 n=10+10)
Gin_GPlusStatic     62.7ns ± 0%    69.5ns ± 2%  +10.90%  (p=0.000 n=10+10)
Gin_GPlusParam      88.6ns ± 3%   105.4ns ± 2%  +18.98%  (p=0.000 n=10+10)
Gin_GPlus2Params     112ns ± 2%     138ns ± 1%  +23.81%  (p=0.000 n=10+10)
Gin_GPlusAll        1.14µs ± 1%    1.38µs ± 1%  +20.95%  (p=0.000 n=10+9)
Gin_ParseStatic     59.2ns ± 0%    67.1ns ± 2%  +13.25%  (p=0.000 n=10+10)
Gin_ParseParam      70.6ns ± 4%    75.7ns ± 2%   +7.22%  (p=0.000 n=10+10)
Gin_Parse2Params    83.3ns ± 1%    90.8ns ± 1%   +8.95%  (p=0.000 n=9+10)
Gin_ParseAll        1.97µs ± 1%    2.27µs ± 1%  +15.50%  (p=0.000 n=10+10)
Gin_StaticAll       18.6µs ± 0%    25.0µs ± 1%  +34.31%  (p=0.000 n=10+10)
```