@qm012 2021-07-01
hi @appleboy I updated the code, and when looking for the contents of the test report, I found that the relevant code submitted with (#2706) maintained two nodes, which would cause the test report to fail to get the expected results, so I made changes and optimizations. By maintaining a node, I controlled the way to find nodes, and made several cross-test reports for comparison when judging test conditions, and got an acceptable final result. Thank you for your help and review.

The benchmarks:

branch| benchmark| running time
:-: | :-: | :-: 
**`master`**  | https://travis-ci.com/github/qm012/go-http-routing-benchmark/builds/231451724 | 634.111s
**`this PR`**  | https://travis-ci.com/github/qm012/go-http-routing-benchmark/builds/231449191 | 633.627s 

The benchstat report:
```shell
name              old time/op    new time/op    delta
Gin_Param           63.5ns ± 4%    62.6ns ± 4%    ~     (p=0.448 n=10+10)
Gin_Param5           112ns ± 6%     114ns ± 3%    ~     (p=0.109 n=10+10)
Gin_Param20          282ns ± 3%     297ns ± 1%  +5.47%  (p=0.000 n=10+10)
Gin_ParamWrite       119ns ± 5%     118ns ± 2%    ~     (p=0.566 n=10+10)
Gin_GithubStatic    81.2ns ± 6%    79.3ns ± 0%    ~     (p=0.085 n=10+10)
Gin_GithubParam      128ns ± 6%     132ns ± 1%  +2.64%  (p=0.006 n=10+10)
Gin_GithubAll       26.2µs ± 3%    27.1µs ± 1%  +3.10%  (p=0.004 n=10+10)
Gin_GPlusStatic     66.0ns ± 3%    63.3ns ± 2%  -4.17%  (p=0.000 n=10+10)
Gin_GPlusParam      91.2ns ± 3%    86.1ns ± 1%  -5.62%  (p=0.000 n=10+10)
Gin_GPlus2Params     117ns ± 5%     110ns ± 1%  -6.32%  (p=0.000 n=9+10)
Gin_GPlusAll        1.23µs ± 7%    1.22µs ± 0%    ~     (p=1.000 n=10+8)
Gin_ParseStatic     62.4ns ± 5%    61.9ns ± 1%    ~     (p=0.218 n=10+10)
Gin_ParseParam      70.2ns ± 6%    68.3ns ± 1%    ~     (p=0.184 n=10+10)
Gin_Parse2Params    82.5ns ± 4%    83.7ns ± 1%  +1.49%  (p=0.014 n=10+9)
Gin_ParseAll        2.01µs ± 3%    2.05µs ± 2%    ~     (p=0.287 n=10+10)
Gin_StaticAll       18.9µs ± 3%    19.4µs ± 1%  +2.93%  (p=0.000 n=10+10)
```