@qm012 2021-07-02
hi @appleboy After a lot of cross-testing and comparison, the latest feedback test report is very exciting.I also tried to match '/' in another way, but the final result was not satisfactory.
[source code](https://github.com/qm012/gin/compare/vtest16...test) 
[benchmark](https://travis-ci.com/github/qm012/go-http-routing-benchmark/builds/231584656)

The latest benchmarks:
branch| benchmark| running time
:-: | :-: | :-: 
**`master`**  | https://travis-ci.com/github/qm012/go-http-routing-benchmark/builds/231451724 | 634.111s
**`this PR`**  | https://travis-ci.com/github/qm012/go-http-routing-benchmark/builds/231589953 | 631.022s

The latest benchstat report:
```shell
name              old time/op    new time/op    delta
Gin_Param           63.5ns ± 4%    60.6ns ± 1%   -4.54%  (p=0.000 n=10+10)
Gin_Param5           112ns ± 6%     103ns ± 1%   -7.63%  (p=0.000 n=10+9)
Gin_Param20          282ns ± 3%     259ns ± 3%   -7.91%  (p=0.000 n=10+10)
Gin_ParamWrite       119ns ± 5%     117ns ± 3%     ~     (p=0.160 n=10+10)
Gin_GithubStatic    81.2ns ± 6%    71.9ns ± 2%  -11.36%  (p=0.000 n=10+10)
Gin_GithubParam      128ns ± 6%     119ns ± 1%   -7.16%  (p=0.000 n=10+10)
Gin_GithubAll       26.2µs ± 3%    25.3µs ± 1%   -3.41%  (p=0.000 n=10+10)
Gin_GPlusStatic     66.0ns ± 3%    58.2ns ± 2%  -11.86%  (p=0.000 n=10+10)
Gin_GPlusParam      91.2ns ± 3%    76.0ns ± 2%  -16.70%  (p=0.000 n=10+8)
Gin_GPlus2Params     117ns ± 5%      95ns ± 1%  -19.17%  (p=0.000 n=9+8)
Gin_GPlusAll        1.23µs ± 7%    1.09µs ± 2%  -11.36%  (p=0.000 n=10+10)
Gin_ParseStatic     62.4ns ± 5%    62.0ns ± 2%     ~     (p=0.280 n=10+10)
Gin_ParseParam      70.2ns ± 6%    69.0ns ± 4%     ~     (p=0.218 n=10+10)
Gin_Parse2Params    82.5ns ± 4%    83.0ns ± 2%     ~     (p=0.271 n=10+10)
Gin_ParseAll        2.01µs ± 3%    1.94µs ± 2%   -3.86%  (p=0.008 n=10+10)
Gin_StaticAll       18.9µs ± 3%    18.4µs ± 1%   -2.59%  (p=0.005 n=10+10)
```