@qm012 2021-07-05
> > hi @appleboy After a lot of cross-testing and comparison, the latest feedback test report is very exciting.I also tried to match '/' in another way, but the final result was not satisfactory.
> > [source code](https://github.com/qm012/gin/compare/vtest16...test)
> > [benchmark](https://travis-ci.com/github/qm012/go-http-routing-benchmark/builds/231584656)
> > The latest benchmarks:
> > branch	benchmark	running time
> > **`master`**	https://travis-ci.com/github/qm012/go-http-routing-benchmark/builds/231451724	634.111s
> > **`this PR`**	https://travis-ci.com/github/qm012/go-http-routing-benchmark/builds/231589953	631.022s
> > The latest benchstat report:
> > ```shell
> > name              old time/op    new time/op    delta
> > Gin_Param           63.5ns ± 4%    60.6ns ± 1%   -4.54%  (p=0.000 n=10+10)
> > Gin_Param5           112ns ± 6%     103ns ± 1%   -7.63%  (p=0.000 n=10+9)
> > Gin_Param20          282ns ± 3%     259ns ± 3%   -7.91%  (p=0.000 n=10+10)
> > Gin_ParamWrite       119ns ± 5%     117ns ± 3%     ~     (p=0.160 n=10+10)
> > Gin_GithubStatic    81.2ns ± 6%    71.9ns ± 2%  -11.36%  (p=0.000 n=10+10)
> > Gin_GithubParam      128ns ± 6%     119ns ± 1%   -7.16%  (p=0.000 n=10+10)
> > Gin_GithubAll       26.2µs ± 3%    25.3µs ± 1%   -3.41%  (p=0.000 n=10+10)
> > Gin_GPlusStatic     66.0ns ± 3%    58.2ns ± 2%  -11.86%  (p=0.000 n=10+10)
> > Gin_GPlusParam      91.2ns ± 3%    76.0ns ± 2%  -16.70%  (p=0.000 n=10+8)
> > Gin_GPlus2Params     117ns ± 5%      95ns ± 1%  -19.17%  (p=0.000 n=9+8)
> > Gin_GPlusAll        1.23µs ± 7%    1.09µs ± 2%  -11.36%  (p=0.000 n=10+10)
> > Gin_ParseStatic     62.4ns ± 5%    62.0ns ± 2%     ~     (p=0.280 n=10+10)
> > Gin_ParseParam      70.2ns ± 6%    69.0ns ± 4%     ~     (p=0.218 n=10+10)
> > Gin_Parse2Params    82.5ns ± 4%    83.0ns ± 2%     ~     (p=0.271 n=10+10)
> > Gin_ParseAll        2.01µs ± 3%    1.94µs ± 2%   -3.86%  (p=0.008 n=10+10)
> > Gin_StaticAll       18.9µs ± 3%    18.4µs ± 1%   -2.59%  (p=0.005 n=10+10)
> > ```
> 
> @qm012 thanks a lot!
> can you run benchmark test using the latest pr code again? thanks!

1. The previous test report was better because I removed the 'value.tsr' match, but that would have caused a TSR error, and I have restored that piece of code
https://github.com/gin-gonic/gin/blob/9c27053243cb24ecc90d01c8ff379bd98fed9c8e/tree.go#L576

2. There will be a derived bug matching to 'Level 1 Router' parameter in the current bug, and one bit will be intercepted. I wanted to open a new PR to fix it before, but I didn't put it in this PR. Can I deal with it like this?

The latest benchmarks:

branch| benchmark| running time
:-: | :-: | :-: 
**`master`**  | https://travis-ci.com/github/qm012/go-http-routing-benchmark/builds/231451724 | 634.111s
**`this PR`** | https://travis-ci.com/github/qm012/go-http-routing-benchmark/builds/231780043 | 634.919s

The latest  benchstat report:
```shell
name              old time/op    new time/op    delta
Gin_Param           63.5ns ± 4%    60.4ns ± 1%   -4.87%  (p=0.000 n=10+9)
Gin_Param5           112ns ± 6%     119ns ± 4%   +6.74%  (p=0.000 n=10+10)
Gin_Param20          282ns ± 3%     310ns ± 2%  +10.09%  (p=0.000 n=10+10)
Gin_ParamWrite       119ns ± 5%     115ns ± 1%   -3.23%  (p=0.003 n=10+10)
Gin_GithubStatic    81.2ns ± 6%    80.6ns ± 2%     ~     (p=0.579 n=10+10)
Gin_GithubParam      128ns ± 6%     138ns ± 2%   +7.66%  (p=0.000 n=10+10)
Gin_GithubAll       26.2µs ± 3%    27.8µs ± 0%   +5.99%  (p=0.000 n=10+10)
Gin_GPlusStatic     66.0ns ± 3%    61.6ns ± 2%   -6.70%  (p=0.000 n=10+10)
Gin_GPlusParam      91.2ns ± 3%    91.7ns ± 0%     ~     (p=0.500 n=10+8)
Gin_GPlus2Params     117ns ± 5%     119ns ± 0%   +1.30%  (p=0.003 n=9+10)
Gin_GPlusAll        1.23µs ± 7%    1.20µs ± 1%     ~     (p=0.671 n=10+10)
Gin_ParseStatic     62.4ns ± 5%    60.1ns ± 1%   -3.60%  (p=0.001 n=10+10)
Gin_ParseParam      70.2ns ± 6%    66.8ns ± 1%   -4.83%  (p=0.000 n=10+9)
Gin_Parse2Params    82.5ns ± 4%    84.3ns ± 0%   +2.29%  (p=0.001 n=10+10)
Gin_ParseAll        2.01µs ± 3%    1.99µs ± 1%     ~     (p=0.469 n=10+10)
Gin_StaticAll       18.9µs ± 3%    19.7µs ± 0%   +4.49%  (p=0.000 n=10+10)
```